Move dashboard check from fmt to check
